Development commit
* chore: add deps for deployment

* fix: change branch name on deploy.yml

* fix: change branch name

* chore: change deploy path

* fix: change php path on deployer

* fix: commented header

* fix: add optional httacces

* fix: change deployment path

* fix admin seeder to login account (#1)

Fix Login

* Fix/dashboard client (#2)

* responsive untuk hero,  about, porto

* angel/feat/masih modif responsif nav, hero, about, porto, roles

* angel/fix/responsif nav, hero, about, porto

* angel/fix/bug portfolio

* angel/feat/responsif for tools, founder, and faqs

* angel/fix/responsif testimonial dan tampilan contact dikit

* angel/fix/tampilan testimoni, contact, and footer responsif

* fix : admin table

* fix: ganti map footer

* fix: navbar responsive

* fix : tools client

* fix: change about text

* fix: delete founder name

* fix: style css and responsive css

* Fix/admin seeder kahasolusi (#5)

* fix admin seeder to login account

* feat: add admin seed

* fix/Seeder Admin

* fix:roles (#6)

* Fix/styling css (#7)

* fix:roles

* fix: styling css add style-2.css

* fix: link css

* fix/admin-sidebar (#8)

admin sidebar menjadi lebih bagus, pengoptimalan tampilan

* Features/add tool table column (#10)

* fix:roles

* fix: styling css add style-2.css

* fix: link css

* fix: tools dashboard

* fix: tools admin

* fix: roles padding area

* fix: company structure modul

* feat: add new route

* fix:fonts admin

* Memberikan Komentar pada deploy yml

* fix: add gitignore storage

* fix: testimonials responsif dashboard

* fix:responsif-testimonial-dashboard

* fix:portfolios-card-dashboard

* fix:footer-responsif-dashboard

* feat:version(1.2)-css-root-DashboardAdmin

* fix:route welcome

* feature: add dropdown sorting portfolios

* fix: descending first

* fix:logo in footer

* patch: delete login footer

* patch: update table testimonials

* feature: menambah tamilan tahun di card porto

* fix: small bug pop up

* edit admin seeder

* fix: admin seeder with faker

* Membenarkan Factory untuk seeder

* Melebarkan footer

* fix: form tools

* feature: add dinamic testimonials

* fix: hero contact

* feature: add contact feature

* fix&feat: layout dan asset gambar hero

* fix:ubah asset dan memperbaiki responsif about

* fix:responsif portfolio dan title di see-more

* fix:tampilan footer deskop dan mobile

* feat&fix: version dan css

* fix:hapus gambar yang tidak digunakan

* versioning login and admin

* fix: navbar see more

* fix: descending controller fixed

* fix:faq admin delete button

* fix: delete btn faq destroy

* fix: redirect update data to index table

* fix : FAQ close popup

* fix: position faqs roles and tools

* fix&feat:gradient-background-and-tools-responsif

* fix: responsive and add 500 blade

* fix: tools client

* fix:controller validator

* fix: controller validator

* fix/portfolio

* fix/portfolios-responsif

* fix:role controller show button true

* fix: errors page

* fix: tools-lg-size

* fix:email

---------

+++ app/Http/Controllers/ToolController.php
@@ -5,8 +5,8 @@ namespace App\Http\Controllers;
 use App\Models\Tool;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
-use Intervention\Image\Facades\Image;
 use Illuminate\Support\Facades\Storage;
+use Intervention\Image\Facades\Image;
 
 class ToolController extends Controller
 {
@@ -16,8 +16,9 @@ class ToolController extends Controller
     public function index()
     {
         if (Auth::guard('admin')->check()) {
-            $tools = Tool::orderBy('created_at', 'DESC')->paginate(8);
+            $tools = Tool::orderBy('updated_at', 'DESC')->get();
             $showButton = true;
+
             return view('admin.tools.index', compact('tools', 'showButton'));
         }
     }
@@ -39,7 +40,9 @@ class ToolController extends Controller
     {
         if (Auth::guard('admin')->check()) {
             $request->validate([
-                'tool_image_url'  => 'required|image|mimes:jpeg,jpg,png|max:2048'
+                'tool_title' => 'required|string|max:15',
+                'tool_desc' => 'required|string|max:15',
+                'tool_image_url' => 'required|image|mimes:jpeg,jpg,png|max:2048',
             ]);
 
             // ini untuk mendapatkan original filename
@@ -48,20 +51,22 @@ class ToolController extends Controller
             // ini untuk mendapatkan extension originalnya
             $originalExtension = $request->file('tool_image_url')->getClientOriginalExtension();
             // ini adalah nama file yang akan disimpan ke database
-            $savedFileName = $filename . '_' . time() . '.' . $originalExtension;
+            $savedFileName = $filename.'_'.time().'.'.$originalExtension;
             // ini adalah path tempat menaruh foto di dalam foldernya di laravel
-            $path = storage_path('app/public/tools/' . $savedFileName);
+            $path = storage_path('app/public/tools/'.$savedFileName);
             $photoResized = Image::make($request->file('tool_image_url'));
-            $photoResized->fit(258,110)->save($path);
+            $photoResized->fit(100, 100)->save($path);
             // ini untuk create datanya
             tool::create([
-                'tool_image_url'  => $savedFileName,
-                'admin_id'        => Auth::guard('admin')->id(),
+                'tool_title' => $request->tool_title,
+                'tool_desc' => $request->tool_desc,
+                'tool_image_url' => $savedFileName,
+                'admin_id' => Auth::guard('admin')->id(),
             ]);
 
-            return redirect()->route('tools.create')->with(['message'   =>  'Tool Berhasil Ditambahkan!']);
+            return redirect()->route('tools.create')->with(['message' => 'Tool Berhasil Ditambahkan!']);
         }
-        
+
     }
 
     /**
@@ -71,6 +76,7 @@ class ToolController extends Controller
     {
         if (Auth::guard('admin')->check()) {
             $tool = Tool::findOrFail($id);
+
             return view('admin.tools.edit', compact('tool'));
         }
     }
@@ -82,34 +88,46 @@ class ToolController extends Controller
     {
         if (Auth::guard('admin')->check()) {
             $request->validate([
-                'tool_image_url'  => 'required|image|mimes:jpeg,jpg,png|max:2048'
+                'tool_title' => 'required|string|max:15',
+                'tool_desc' => 'required|string|max:15',
+                'tool_image_url' => 'nullable|image|mimes:jpeg,jpg,png|max:2048',
             ]);
             $tool = Tool::findOrFail($id);
-    
-            // ini untuk mendapatkan original filename
-            $originalFileName = $request->file('tool_image_url')->getClientOriginalName();
-            $filename = pathinfo($originalFileName, PATHINFO_FILENAME);
-            // ini untuk mendapatkan extension originalnya
-            $originalExtension = $request->file('tool_image_url')->getClientOriginalExtension();
-            // ini adalah nama file yang akan disimpan ke database
-            $savedFileName = $filename . '_' . time() . '.' . $originalExtension;
-            // ini adalah path tempat menaruh foto di dalam foldernya di laravel
-            $path = storage_path('app/public/tools/' . $savedFileName);
-            $photoResized = Image::make($request->file('tool_image_url'));
-            $photoResized->fit(150,150)->save($path);
 
-            // Menghapus data image
-            $imagePath = 'public/tools/'.$tool->tool_image_url;
-            if (Storage::exists($imagePath)) {
-                Storage::delete($imagePath);
+            // percabangan untuk mengecek apakah di request itu punya image atau tidak
+            if ($request->hasFile('tool_image_url')) {
+                // ini untuk mendapatkan original filename
+                $originalFileName = $request->file('tool_image_url')->getClientOriginalName();
+                $filename = pathinfo($originalFileName, PATHINFO_FILENAME);
+                // ini untuk mendapatkan extension originalnya
+                $originalExtension = $request->file('tool_image_url')->getClientOriginalExtension();
+                // ini adalah nama file yang akan disimpan ke database
+                $savedFileName = $filename.'_'.time().'.'.$originalExtension;
+                // ini adalah path tempat menaruh foto di dalam foldernya di laravel
+                $path = storage_path('app/public/tools/'.$savedFileName);
+                $photoResized = Image::make($request->file('tool_image_url'));
+                $photoResized->fit(150, 150)->save($path);
+
+                // Menghapus data image
+                $imagePath = 'public/tools/'.$tool->tool_image_url;
+                if (Storage::exists($imagePath)) {
+                    Storage::delete($imagePath);
+                }
+                // ini untuk mengupdate datanya
+                $tool->update([
+                    'tool_title' => $request->tool_title,
+                    'tool_desc' => $request->tool_desc,
+                    'tool_image_url' => $savedFileName,
+                ]);
+            } else {
+                $tool->update([
+                    'tool_title' => $request->tool_title,
+                    'tool_desc' => $request->tool_desc,
+                ]);
             }
-            // ini untuk mengupdate datanya
-            $tool->update([
-                'tool_image_url'  => $savedFileName,
-            ]);
 
-            //redirect to new edit form
-            return redirect()->route('tools.edit', $tool->tool_id)->with(['message' => 'Tool Berhasil Diubah!']);
+            //redirect to new index table
+            return redirect()->route('tools.index', $tool->tool_id)->with(['message' => 'Tool Berhasil Diubah!']);
         }
     }
 
@@ -122,15 +140,15 @@ class ToolController extends Controller
             // tool objek
             $tool = Tool::findOrFail($id);
             // image path
-            $imagePath = 'public/tools/' . $tool->tool_image_url;
-    
+            $imagePath = 'public/tools/'.$tool->tool_image_url;
+
             // check if image exist
             if (Storage::exists($imagePath)) {
                 Storage::delete($imagePath);
             }
-    
+
             $tool->delete();
-    
+
             return redirect()->route('tools.index')->with(['message' => 'Data Berhasil Dihapus!']);
         }
     }
